Users sidebar should show a highlighted buddy list (Frontend).
This adds support for displaying a buddy list and adding and
removing users to it. The buddy list also displays users in
alphabetically increasing order. For moving users to and from
buddy list, user handlebar is re-rendered at the appropriate place.
Frontend node test has been added which checks if the movement of
buddy user is displayed correctly.

Fixes: #236.