Users sidebar should show a highlighted buddy list (Backend).
This adds support for displaying a buddy list and adding and
removing users to it. The buddy list also displays users in
alphabetically increasing order. For moving users to and from
buddy list, user handlebar is re-rendered at the appropriate place.
A new endpoint has been created which manipulates the buddy list and
a backend test for the same has also been added.

Fixes: #236.